refactor: workspace loader page
get rid of the FactoryLoader page and WorkspaceLoader page in favour of using a single loader page
