+++ packages/dashboard-frontend/src/containers/Loader/Factory/Steps/Apply/Resources/index.tsx
@@ -23,7 +23,7 @@ import * as DevfileRegistriesStore from '../../../../../../store/DevfileRegistri
 import { DisposableCollection } from '../../../../../../services/helpers/disposable';
 import { selectAllWorkspaces } from '../../../../../../store/Workspaces/selectors';
 import { delay } from '../../../../../../services/helpers/delay';
-import { FactoryLoaderPage } from '../../../../../../pages/Loader/Factory';
+import { LoaderPage } from '../../../../../../pages/Loader';
 import { selectDefaultNamespace } from '../../../../../../store/InfrastructureNamespaces/selectors';
 import {
   selectFactoryResolver,
@@ -233,11 +233,12 @@ class StepApplyResources extends AbstractLoaderStep<Props, State> {
     const alertItem = this.getAlertItem(lastError);
 
     return (
-      <FactoryLoaderPage
+      <LoaderPage
         alertItem={alertItem}
         currentStepId={currentStepId}
         steps={steps}
         tabParam={tabParam}
+        workspace={undefined}
         onTabChange={tab => this.handleTabChange(tab)}
       />
     );
