+++ packages/dashboard-frontend/src/containers/Loader/Factory/Steps/Fetch/Devfile/index.tsx
@@ -20,7 +20,7 @@ import * as FactoryResolverStore from '../../../../../../store/FactoryResolver';
 import { DisposableCollection } from '../../../../../../services/helpers/disposable';
 import { selectAllWorkspaces } from '../../../../../../store/Workspaces/selectors';
 import { delay } from '../../../../../../services/helpers/delay';
-import { FactoryLoaderPage } from '../../../../../../pages/Loader/Factory';
+import { LoaderPage } from '../../../../../../pages/Loader';
 import { getEnvironment, isDevEnvironment } from '../../../../../../services/helpers/environment';
 import SessionStorageService, {
   SessionStorageKey,
@@ -359,11 +359,12 @@ class StepFetchDevfile extends AbstractLoaderStep<Props, State> {
     const alertItem = this.getAlertItem(lastError);
 
     return (
-      <FactoryLoaderPage
+      <LoaderPage
         alertItem={alertItem}
         currentStepId={currentStepId}
         steps={steps}
         tabParam={tabParam}
+        workspace={undefined}
         onTabChange={tab => this.handleTabChange(tab)}
       />
     );
