+++ packages/dashboard-frontend/src/containers/Loader/Factory/Steps/Initialize/index.tsx
@@ -21,7 +21,7 @@ import { selectInfrastructureNamespaces } from '../../../../../store/Infrastruct
 import { DisposableCollection } from '../../../../../services/helpers/disposable';
 import { delay } from '../../../../../services/helpers/delay';
 import { ROUTE } from '../../../../../Routes/routes';
-import { FactoryLoaderPage } from '../../../../../pages/Loader/Factory';
+import { LoaderPage } from '../../../../../pages/Loader';
 import { FactoryParams, PoliciesCreate } from '../../types';
 import { MIN_STEP_DURATION_MS } from '../../../const';
 import buildFactoryParams from '../../buildFactoryParams';
@@ -169,11 +169,12 @@ class StepInitialize extends AbstractLoaderStep<Props, State> {
     const alertItem = this.getAlertItem(lastError);
 
     return (
-      <FactoryLoaderPage
+      <LoaderPage
         alertItem={alertItem}
         currentStepId={currentStepId}
         steps={steps}
         tabParam={tabParam}
+        workspace={undefined}
         onTabChange={tab => this.handleTabChange(tab)}
       />
     );
