--- packages/dashboard-frontend/src/pages/Loader/Workspace/index.tsx
@@ -1,67 +1,0 @@
-/*
- * Copyright (c) 2018-2023 Red Hat, Inc.
- * This program and the accompanying materials are made
- * available under the terms of the Eclipse Public License 2.0
- * which is available at https://www.eclipse.org/legal/epl-2.0/
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *   Red Hat, Inc. - initial API and implementation
- */
-
-import React from 'react';
-import { LoaderStep } from '../../../components/Loader/Step';
-import { AlertItem, LoaderTab } from '../../../services/helpers/types';
-import { Workspace } from '../../../services/workspace-adapter';
-import { CommonLoaderPage } from '../Common';
-
-export type Props = {
-  alertItem: AlertItem | undefined;
-  currentStepId: number;
-  steps: LoaderStep[];
-  tabParam: string | undefined;
-  workspace: Workspace | undefined;
-  onTabChange: (tab: string) => void;
-};
-
-export type State = {
-  activeTabKey: LoaderTab;
-};
-
-export default class WorkspaceLoaderPage extends React.PureComponent<Props, State> {
-  constructor(props: Props) {
-    super(props);
-
-    const { tabParam } = this.props;
-    const activeTabKey = tabParam && LoaderTab[tabParam] ? LoaderTab[tabParam] : LoaderTab.Progress;
-
-    this.state = {
-      activeTabKey,
-    };
-  }
-
-  private handleTabChange(tabKey: LoaderTab): void {
-    this.setState({
-      activeTabKey: tabKey,
-    });
-    const tabName = LoaderTab[tabKey];
-    this.props.onTabChange(tabName);
-  }
-
-  render(): React.ReactNode {
-    const { alertItem, currentStepId, steps, workspace } = this.props;
-    const { activeTabKey } = this.state;
-
-    return (
-      <CommonLoaderPage
-        activeTabKey={activeTabKey}
-        alertItem={alertItem}
-        currentStepId={currentStepId}
-        steps={steps}
-        workspace={workspace}
-        onTabChange={tabKey => this.handleTabChange(tabKey)}
-      />
-    );
-  }
-}
